Fix bugs from review
