+++ index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/jpg" href="/Adt_icon.jpeg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-    <script src="/prebid10.10.0.js"></script>
+<!--    <script src="/prebid10.10.0.js"></script>-->
 
       <title>News site</title>
   </head>
